requirements/base.txt
Django==2.1.1
pytz==2018.5
six==1.11.0
requests==2.18.4

django-debug-toolbar==1.10.1

psycopg2==2.7.5

